renamed the bundler and package names
