+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.cascadebikeclub",
+    package = "org.cascade",
 )
 
 android_resource(
     name = "res",
-    package = "com.cascadebikeclub",
+    package = "org.cascade",
     res = "src/main/res",
 )
