+++ android/app/src/main/java/org/cascade/MainActivity.java
@@ -1,4 +1,4 @@
-package com.cascadebikeclub;
+package org.cascade;
 
 import com.facebook.react.ReactActivity;
 import android.os.Bundle;
