+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "CascadeBikeClub",
-  "displayName": "CascadeBikeClub"
+  "name": "Cascade",
+  "displayName": "Cascade"
 }
