+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cascadebikeclub",
+  "name": "cascade",
   "version": "0.0.1",
   "private": true,
   "scripts": {
